fix: UI Interface for pagination move to correct file
