--- ui/src/app/workflows/pagination.ts
@@ -1,7 +1,0 @@
-export interface WorkflowsPagination {
-    wfOffset: string;
-    archivedOffset: string;
-    nextWfOffset: string;
-    nextArchivedOffset: string;
-    limit: number;
-}
